chore: Standardize repository maintenance
The package needs a consistent open-source maintenance baseline. This
aligns repository policy, documentation, pinned automation, release and
security workflows, and executable contract checks while retaining the
queue-specific integration gate and bounded benchmark iteration count.
No runtime delivery semantics are changed.

+++ rabbitmq/worker_test.go
@@ -242,7 +242,7 @@ func TestWorkerRunQueueRequestAndShutdown(t *testing.T) {
 	assert.ErrorIs(t, worker.Run(context.Background(), &message), expectedRun)
 	require.NoError(t, worker.Queue(&message))
 	assert.Equal(t, message.Bytes(), channel.published.Body)
-	assert.Equal(t, uint8(amqp.Persistent), channel.published.DeliveryMode)
+	assert.Equal(t, amqp.Persistent, channel.published.DeliveryMode)
 	deadline, ok := channel.publishContext.Deadline()
 	assert.True(t, ok)
 	assert.WithinDuration(t, time.Now().Add(5*time.Second), deadline, time.Second)
